+++ kaye/cli/cli_skill/__init__.py
@@ -1,0 +1,2 @@
+def convert_display_name2skill_name(display_name):
+    return display_name.replace(" ", "-").lower()