+++ tests/cli/c/c/others/cli-c-c-bp-number-unit_test.py
@@ -9,9 +9,9 @@ creation of ``Numerical Values with Units.md``
 import pytest
 
 from tests.cli import MD_FILENAME2SKILL_NAME
-from tests.cli.c.c import (
-    assert_rule_file_basic_format,
-    split_rule_file_basic_format,
+from tests.cli import (
+    assert_frontmatter_md_file_basic_structure,
+    split_frontmatter_md_file,
     assert_header_line_always_apply,
 )
 
@@ -36,12 +36,12 @@ def testee(testee_path):
 
 @pytest.fixture(scope="session")
 def testee_header(testee):
-    return split_rule_file_basic_format(testee)[0]
+    return split_frontmatter_md_file(testee)[0]
 
 
 @pytest.fixture(scope="session")
 def testee_content(testee):
-    return split_rule_file_basic_format(testee)[1]
+    return split_frontmatter_md_file(testee)[1]
 
 
 # Pytest unit tests  ###########################################################
@@ -56,7 +56,7 @@ class TestBasic:  # ============================================================
         assert testee_path.is_file()
 
     def test_structure(_, testee):
-        assert assert_rule_file_basic_format(testee)
+        assert assert_frontmatter_md_file_basic_structure(testee)
 
 
 class TestHeader:  # ===========================================================